chore(sidecar): make spark etl to write to cockroachdb
