+++ _ops/dev/env.dev-sample.sh
@@ -27,6 +27,7 @@ ZOOKEEPER_CONNECT=kafka:2181
 KAFKA_BROKERS=kafka:9092
 KAFKA_TOPICS=topic1
 KAFKA_GROUP_ID=maiev_dev
+KAFKA_MAX_OFFSET_PER_TRIGGER=
 
 # TwitterKafkaProducer
 ACCESS_TOKEN=
