+++ src/main/scala/org/leletan/maiev/lib/KafkaTopicData.scala
@@ -1,0 +1,6 @@
+package org.leletan.maiev.lib
+
+/**
+ * Created by jiale.tan on 12/23/17.
+ */
+case class KafkaTopicData (topic: String, value: String)
